backend/ai-service/requirements.txt
fastapi
uvicorn
python-dotenv
opencv-python-headless
tflite-runtime
python-multipart
numpy
PyJWT
requests